Problems with storage again

+++ src/solver/ODE_typedefs.h
@@ -14,7 +14,7 @@
 
 /** @name Default data types for solver */
 /**@{*/
-typedef double	ODE_R;  /**< type for continuous variables */
+typedef long double	ODE_R;  /**< type for continuous variables */
 typedef unsigned int	ODE_uint;	/**< type for unsigned in, (mainly flags) */
 /**@}*/
